refactor(dto): Move dto package to pkg root
