+++ pkg/handlers/auth.go
@@ -15,10 +15,10 @@ import (
 	"github.com/kptm-tools/core-service/pkg/auth"
 	"github.com/kptm-tools/core-service/pkg/config"
 	"github.com/kptm-tools/core-service/pkg/domain"
+	"github.com/kptm-tools/core-service/pkg/dto"
 	"github.com/kptm-tools/core-service/pkg/interfaces"
 	"github.com/kptm-tools/core-service/pkg/middleware"
 	"github.com/kptm-tools/core-service/pkg/services"
-	"github.com/kptm-tools/core-service/pkg/ws/report/dto"
 )
 
 type AuthHandlers struct {
