+++ pkg/handlers/host.go
@@ -7,9 +7,9 @@ import (
 	"net/http"
 	"strconv"
 
+	"github.com/kptm-tools/core-service/pkg/dto"
 	"github.com/kptm-tools/core-service/pkg/middleware"
 	"github.com/kptm-tools/core-service/pkg/services"
-	"github.com/kptm-tools/core-service/pkg/ws/report/dto"
 
 	"github.com/kptm-tools/core-service/pkg/api"
 	"github.com/kptm-tools/core-service/pkg/domain"
