+++ pkg/handlers/scanschedule.go
@@ -9,9 +9,9 @@ import (
 	"github.com/google/uuid"
 	cmmn "github.com/kptm-tools/common/common/pkg/events"
 	"github.com/kptm-tools/core-service/pkg/api"
+	"github.com/kptm-tools/core-service/pkg/dto"
 	"github.com/kptm-tools/core-service/pkg/interfaces"
 	"github.com/kptm-tools/core-service/pkg/middleware"
-	"github.com/kptm-tools/core-service/pkg/ws/report/dto"
 )
 
 type ScanScheduleHandlers struct {
