+++ pkg/handlers/vulnerability.go
@@ -9,8 +9,8 @@ import (
 
 	"github.com/kptm-tools/core-service/pkg/api"
 	"github.com/kptm-tools/core-service/pkg/customerrors"
+	"github.com/kptm-tools/core-service/pkg/dto"
 	"github.com/kptm-tools/core-service/pkg/interfaces"
-	"github.com/kptm-tools/core-service/pkg/ws/report/dto"
 )
 
 type VulnerabilityHandlers struct {
