+++ pkg/ws/report/handlers/apply_vectors.go
@@ -5,7 +5,7 @@ import (
 	"log/slog"
 
 	"github.com/kptm-tools/core-service/pkg/customerrors"
-	"github.com/kptm-tools/core-service/pkg/ws/report/dto"
+	"github.com/kptm-tools/core-service/pkg/dto"
 	"github.com/kptm-tools/core-service/pkg/ws/report/reportutils"
 
 	"github.com/kptm-tools/core-service/pkg/interfaces"
