+++ pkg/ws/report/handlers/initial_request.go
@@ -6,9 +6,9 @@ import (
 
 	"github.com/google/uuid"
 	"github.com/kptm-tools/core-service/pkg/customerrors"
+	"github.com/kptm-tools/core-service/pkg/dto"
 	"github.com/kptm-tools/core-service/pkg/interfaces"
 	"github.com/kptm-tools/core-service/pkg/ws/common"
-	"github.com/kptm-tools/core-service/pkg/ws/report/dto"
 	"github.com/kptm-tools/core-service/pkg/ws/report/reportutils"
 )
