+++ pkg/ws/report/handlers/vector_update.go
@@ -6,7 +6,7 @@ import (
 	"log/slog"
 
 	"github.com/kptm-tools/core-service/pkg/customerrors"
-	"github.com/kptm-tools/core-service/pkg/ws/report/dto"
+	"github.com/kptm-tools/core-service/pkg/dto"
 	"github.com/kptm-tools/core-service/pkg/ws/report/reportutils"
 
 	"github.com/kptm-tools/common/common/pkg/enums"
